tratando erro

+++ src/pages/Repositories/index.js
@@ -1,23 +1,31 @@
 import React, { useEffect, useState } from "react";
 import * as S from './styled';
+import { useNavigate } from 'react-router-dom';
 
 export default function Repositories() {
-    const [ repositories, setRepositories ] = useState([]);
+    const navigate = useNavigate();
+    const [repositories, setRepositories] = useState([]);
 
     useEffect(() => {
         let repositoriesName = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
-        setRepositories(repositoriesName);
-        localStorage.clear();
-    }, []);
+        console.log(repositoriesName)
+        if (repositoriesName != null) {
+            repositoriesName = JSON.parse(repositoriesName);
+            setRepositories(repositoriesName);
+            localStorage.clear();
+        } else {
+            navigate('/')
+        }
+
+    }, [navigate]);
 
     return (
         <S.Container>
             <S.Title>Repositórios</S.Title>
             <S.List>
-                { repositories.map(repository => {
+                {repositories.map(repository => {
                     return (
-                        <S.ListItem>Repositório: { repository }</S.ListItem>
+                        <S.ListItem>Repositório: {repository}</S.ListItem>
                     )
                 })}
             </S.List>
